remove unnecessary utils
