+++ core_chat_data/src/androidMain/kotlin/net/thechance/mena/core_chat/data/di/dataProviderModule.android.kt
@@ -1,10 +1,12 @@
 package net.thechance.mena.core_chat.data.di
 
+import android.content.Context
 import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.Preferences
 import com.bilalazzam.contacts_provider.ContactsProvider
 import com.bilalazzam.contacts_provider.ContactsProviderFactory
-import net.thechance.mena.core_chat.data.utils.createDataStore
+import net.thechance.mena.core_chat.data.contacts.utils.createDataStore
+import net.thechance.mena.core_chat.data.contacts.utils.dataStoreName
 import org.koin.core.scope.Scope
 
 actual fun Scope.createContactsProvider(): ContactsProvider {
@@ -12,5 +14,8 @@ actual fun Scope.createContactsProvider(): ContactsProvider {
 }
 
 actual fun Scope.createSettingsDataStore(): DataStore<Preferences>{
-    return createDataStore(context = this.get())
+    val context = this.get<Context>()
+    return createDataStore {
+        context.filesDir.resolve(dataStoreName).absolutePath
+    }
 }
