--- core_chat_data/src/iosMain/kotlin/net/thechance/mena/core_chat/data/utils/CreateDataStore.ios.kt
@@ -1,25 +1,0 @@
-package net.thechance.mena.core_chat.data.utils
-
-import androidx.datastore.core.DataStore
-import androidx.datastore.preferences.core.Preferences
-import kotlinx.cinterop.ExperimentalForeignApi
-import net.thechance.mena.core_chat.data.contacts.utils.createDataStore
-import net.thechance.mena.core_chat.data.contacts.utils.dataStoreName
-import platform.Foundation.NSDocumentDirectory
-import platform.Foundation.NSUserDomainMask
-import platform.Foundation.*
-
-
-@OptIn(ExperimentalForeignApi::class)
-fun createDataStore(): DataStore<Preferences> {
-    return createDataStore {
-        val documentDirectory: NSURL? = NSFileManager.defaultManager.URLForDirectory(
-            directory = NSDocumentDirectory,
-            inDomain = NSUserDomainMask,
-            appropriateForURL = null,
-            create = false,
-            error = null,
-        )
-        requireNotNull(documentDirectory).path + "/$dataStoreName"
-    }
-}
