+++ core_chat_domain/build.gradle.kts
@@ -9,7 +9,6 @@ kotlin {
 
     sourceSets {
         commonMain.dependencies {
-            implementation(libs.kotlinx.coroutines.core)
 
         }
     }
